pyproject.toml
[project]
name = "julian-gregory"
version = "0.1.0"
description = "The Calendar Assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.15.0,<2.0.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "pytest>=9.0.1",
]

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = [
    "ignore::DeprecationWarning:google.adk.runners"
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
